+++ lib/firebase_auth_simplify.dart
@@ -2,5 +2,7 @@ library firebase_auth_simplify;
 
 export 'src/firebase_auth_provider.dart';
 export 'src/api/base_auth_api.dart';
+
 export 'src/api/firebase_email_auth_api.dart';
 export 'src/api/firebase_google_auth_api.dart';
+export 'src/api/firebase_kakao_auth_api.dart';