+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.0] - 27/Jan/2020
+
+- Supporting Kakao sign in
+
 ## [0.0.1] - 25/Jan/2020
 
-- Initial release
+- Initial release with email, Gmail support